// src/shared/api/generated/schemas/auctionListItemTradingStatusMobile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Торговый статус пользователя в аукционе:
 *
 * * **NotParticipating** — не участвует (1)
 * * **Leading** — лидирует (2)
 * * **Losing** — перебит (3)
 * * **Winner** — победитель (4)
 * * **Confirmed** — подтверждён (5)
 * * **Unknown** — неизвестный статус
 */
export type AuctionListItemTradingStatusMobile = typeof AuctionListItemTradingStatusMobile[keyof typeof AuctionListItemTradingStatusMobile];


export const AuctionListItemTradingStatusMobile = {
  NotParticipating: 'NotParticipating',
  Leading: 'Leading',
  Losing: 'Losing',
  Winner: 'Winner',
  Confirmed: 'Confirmed',
  Unknown: 'Unknown',
} as const;